refactor(web): extract WorkflowRunQueryService + WorkflowRunCancellationService (#167)
Move the three ``/api/workflows/runs`` handlers (list, get single,
cancel) into dedicated services. ``build_run_response`` / ``parse_run_id``
/ ``serialize_run`` stay re-exported under their legacy ``_``-prefixed
names from the router so the ``from srunx.web.routers.workflows import
_build_run_response`` call in tests/transport/test_review_fixes.py keeps
resolving.

Cancellation preserves the same IMMEDIATE-transaction envelope and
race-retry logic against concurrent poller writes as the original
inline handler.
